+++ src/utils.js
@@ -91,50 +91,36 @@ export function getParams(ns, opt, cb) {
 }
 
 const NAME_KEY_SEP = '.';
+const NAME_INDEX_OPEN_SEP = '[';
 
-export function getNameKeyStr(name, key) {
-  if (key) {
-    return `${name}${NAME_KEY_SEP}${key}`;
-  }
-  return name;
-}
+export function getNameIfNested(str) {
+  const keyIndex = str.indexOf(NAME_KEY_SEP);
+  const arrayIndex = str.indexOf(NAME_INDEX_OPEN_SEP);
 
-export function getNameKeyObj(str) {
-  const index = str.indexOf(NAME_KEY_SEP);
-  if (str.indexOf(NAME_KEY_SEP) !== -1) {
-    const name = str.slice(0, index);
-    const key = str.slice(index + NAME_KEY_SEP.length);
+  let index;
+
+  if (keyIndex === -1 && arrayIndex === -1) {
     return {
-      name,
-      key,
+      name: str,
     };
+  } else if (keyIndex === -1) {
+    index = arrayIndex;
+  } else if (arrayIndex === -1) {
+    index = keyIndex;
+  } else {
+    index = Math.min(keyIndex, arrayIndex);
   }
-  return {
-    name: str,
-  };
-}
 
-export function flatFields(fields_, fieldsMeta) {
-  const fields = { ...fields_ };
-  Object.keys(fields).forEach((k) => {
-    if (fieldsMeta[k] && fieldsMeta[k].virtual) {
-      const value = fields[k];
-      // flatten
-      for (const k2 in value) {
-        if (value.hasOwnProperty(k2)) {
-          fields[getNameKeyStr(k, k2)] = value[k2];
-        }
-      }
-      delete fields[k];
-    }
-  });
-  return fields;
+  return {
+    name: str.slice(0, index),
+    isNested: true,
+  };
 }
 
 export function flatFieldNames(names) {
   const ret = {};
   names.forEach((n) => {
-    ret[getNameKeyObj(n).name] = 1;
+    ret[getNameIfNested(n).name] = 1;
   });
   return Object.keys(ret);
 }
@@ -143,9 +129,26 @@ export function clearVirtualField(name, fields, fieldsMeta) {
   if (fieldsMeta[name] && fieldsMeta[name].virtual) {
     /* eslint no-loop-func:0 */
     Object.keys(fields).forEach((ok) => {
-      if (getNameKeyObj(ok).name === name) {
+      if (getNameIfNested(ok).name === name) {
         delete fields[ok];
       }
     });
   }
+}
+
+export function getVirtualPaths(fieldsMeta) {
+  const virtualPaths = {};
+  for (const name in fieldsMeta) {
+    if (fieldsMeta.hasOwnProperty(name)) {
+      const leadingName = fieldsMeta[name].leadingName;
+      if (leadingName && fieldsMeta[leadingName].virtual) {
+        if (leadingName in virtualPaths) {
+          virtualPaths[leadingName].push(name);
+        } else {
+          virtualPaths[leadingName] = [name];
+        }
+      }
+    }
+  }
+  return virtualPaths;
 }